EIP1-492: EIP1-1173: Refactors around RegisterCheckService
