+++ README.md
@@ -1,7 +1,34 @@
-# # eip-ero-register-checker-api
+# eip-ero-register-checker-api
 Spring Boot microservice that is :
 - Consuming and handling SQS messages related to applicant's register checks.
 - Exposing secured REST APIs to allow search and update operations related to applicant's register checks.
+
+# External api dependencies
+- Soft-wire IER api 
+- [eip-ero-management-api](https://github.com/cabinetoffice/eip-ero-management-api)
+
+## Developer Setup
+### Kotlin API Developers
+
+Configure your IDE with the code formatter (ktlint):
+```
+$ ./gradlew ktlintApplyToIdea
+```
+This only needs doing once to setup your IDE with the code styles.
+
+#### Running Tests
+In order to run the tests successfully, you will first need to set the `LOCALSTACK_API_KEY` environment variable (i.e.
+within your .bash_profile or similar). Then run:
+```
+$ ./gradlew check
+```
+This will run the tests and ktlint. (Be warned, ktlint will hurt your feelings!)
+
+#### Building docker images
+```
+$ ./gradlew check bootBuildImage
+```
+This will build a docker image for the Spring Boot application.
 
 ## Running the application
 Either `./gradlew bootRun` or run the class `RegisterCheckerApiApplication`
