--- src/main/kotlin/uk/gov/dluhc/registercheckerapi/service/IerService.kt
@@ -1,11 +1,0 @@
-package uk.gov.dluhc.registercheckerapi.service
-
-import org.springframework.stereotype.Service
-import uk.gov.dluhc.registercheckerapi.client.IerApiClient
-
-@Service
-class IerService(private val ierApiClient: IerApiClient) {
-
-    fun getEroIdentifierForCertificateSerial(certificateSerial: String): String =
-        ierApiClient.getEroIdentifier(certificateSerial).eroId!!
-}
